+++ CommunistPang/src/model/RectangleCollision.java
@@ -3,8 +3,8 @@ package model;
 public class RectangleCollision implements Collision {
 	
 	@Override
-	public boolean collision(float x1, float y1, float width1, float height1, float x2, float y2, float width2,
-			float height2) {
+	public boolean collision(double x1, double y1, double width1, double height1, double x2, double y2, double width2,
+			double height2) {
 		if (((x1 + width1) < x2) || (x1 > (x2 + width2)) || (y1 > (y2 + height2)) || ((y1 + height1) < y2)) return false;
 		else return true;
 	}
